+++ lib/src/culqi_response.dart
@@ -1,20 +1,27 @@
 abstract class CulqiResponse{}
 
-enum ErrorType{
-  InvalidKey,
-  InvalidCard,
+enum CulqiErrorType{
+  ClientInvalidKey,
+  ClientInvalidCard,
   ServerValidationFailed,
+  InvalidBusiness,
+  InvalidCard,
+  ParameterError,
+  ParameterErrorInvalidEmail,
+  ParameterErrorInvalidCvv,
+  ParameterErrorInvalidExpirationMonth,
+  ParameterErrorInvalidExpirationYear,
   ExceptionTrowed
 }
 
 class CulqiError extends CulqiResponse{
 
-  ErrorType _errorType;
+  CulqiErrorType _errorType;
   int _errorCode;
   String _errorMessage;
   Exception _exception;
 
-  CulqiError.fromType(ErrorType errorType, {errorCode=-1, errorMessage='Unknown Error', exception}){
+  CulqiError.fromType(CulqiErrorType errorType, {errorCode=-1, errorMessage='Unknown Error', exception}){
     _errorType = errorType;
     _errorCode = errorCode;
     _errorMessage = errorMessage;
@@ -22,12 +29,41 @@ class CulqiError extends CulqiResponse{
   }
 
   CulqiError.fromJson(Map<String, dynamic> response, int errorCode){
-    _errorType = ErrorType.ServerValidationFailed;
+
+    String errorType = response['type'];
+
+    if(_errorsOnServer.containsKey(errorType)){
+      if(_errorsOnServer[errorType] == CulqiErrorType.ParameterError){
+        String errorSubType = response['code'];
+
+        if(_errorsFromParameter.containsKey(errorSubType)) _errorType = _errorsFromParameter[errorSubType];
+        else _errorType = CulqiErrorType.ParameterError;
+
+      }else{
+        _errorType = CulqiErrorType.ServerValidationFailed;
+      }
+    }else{
+      _errorType = CulqiErrorType.ServerValidationFailed;
+    }
+
     _errorMessage = response['merchant_message'];
     _errorCode = errorCode;
   }
 
-  ErrorType get errorType     => _errorType;
+  Map<String, CulqiErrorType> _errorsOnServer = {
+    'comercio_invalido' : CulqiErrorType.InvalidBusiness,
+    'card_error' : CulqiErrorType.InvalidCard,
+    'parameter_error' : CulqiErrorType.ParameterError,
+  };
+
+  Map<String, CulqiErrorType> _errorsFromParameter = {
+    'invalid_email' : CulqiErrorType.ParameterErrorInvalidEmail,
+    'invalid_cvv' : CulqiErrorType.ParameterErrorInvalidCvv,
+    'invalid_expiration_month' : CulqiErrorType.ParameterErrorInvalidExpirationMonth,
+    'invalid_expiration_year' : CulqiErrorType.ParameterErrorInvalidExpirationYear
+  };
+
+  CulqiErrorType get errorType     => _errorType;
   int get errorCode     => _errorCode;
   String get errorMessage  => _errorMessage;
   Exception get exception => _exception;