+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.5.2
+
+- Added a better error handling (with more specific errors from Culqi).
+- Added some tests to check the specific errors from Culqi
+
 ## 0.5.1
 
 - Fixed some maintenance suggestions from pub.dev